chore: remove meteor.startup from permalink-pinned
