+++ apps/meteor/client/components/message/toolbar/usePermalinkPinned.tsx
@@ -1,0 +1,43 @@
+import { isE2EEMessage } from '@rocket.chat/core-typings';
+import { useToastMessageDispatch } from '@rocket.chat/ui-contexts';
+import { useEffect } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { MessageAction } from '../../../../app/ui-utils/client/lib/MessageAction';
+import { getPermaLink } from '../../../lib/getPermaLink';
+
+export const usePermalinkPinned = () => {
+	const { t } = useTranslation();
+
+	const dispatchToastMessage = useToastMessageDispatch();
+
+	useEffect(() => {
+		MessageAction.addButton({
+			id: 'permalink-pinned',
+			icon: 'permalink',
+			label: 'Copy_link',
+			context: ['pinned'],
+			async action(_, { message }) {
+				try {
+					const permalink = await getPermaLink(message._id);
+					navigator.clipboard.writeText(permalink);
+					dispatchToastMessage({ type: 'success', message: t('Copied') });
+				} catch (e) {
+					dispatchToastMessage({ type: 'error', message: e });
+				}
+			},
+			condition({ subscription }) {
+				return !!subscription;
+			},
+			order: 5,
+			group: 'menu',
+			disabled({ message }) {
+				return isE2EEMessage(message);
+			},
+		});
+
+		return () => {
+			MessageAction.removeButton('permalink-pinned');
+		};
+	}, [dispatchToastMessage, t]);
+};
