+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -2,7 +2,6 @@ import './jumpToMessage';
 import './jumpToPinMessage';
 import './jumpToSearchMessage';
 import './jumpToStarMessage';
-import './permalinkPinned';
 import './permalinkStar';
 import './pinMessage';
 import './starMessage';
